"ads1_m14_a1: 0/4 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS1-assignments/m14/14.1 Dynamic Median Finding/Solution.java
@@ -36,7 +36,7 @@ class Solution{
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		DynamicMedian obj = new DynamicMedian();
-		int inputs = scan.nextInt();
+		long inputs = scan.nextLong();
 		double element = scan.nextDouble();
 		obj.insertAtMin(element);
 		double median = element;
